requirements.txt
streamlit>=1.32.0
pandas>=2.0.0
plotly>=5.18.0
folium>=0.15.0
streamlit-folium>=0.18.0
streamlit-autorefresh>=1.0.1
requests>=2.31.0
python-dotenv>=1.0.0
pyserial>=3.5
